requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
elevenlabs==0.2.26
requests>=2.31.0
